Add script to allow for checking and logging out
Added script to allow the computer to check if its in customer/master mode at boot. This to delete a icon from desktop if in customer mode.
Script also allows logging out. Logging out is one-way and only works if you already is master. Else the logout request will be silently ignored. (So if a customer clicks the logout icon when its not deleted, nothing will happen, as its in customer mode)

No data from client is processed at all, except for the binary flag of if a query string exists or not, this is used to signal for logout or not,